♻️ refactor : constants, hooks, utils 폴더 정리
