+++ src/components/profile/PasswordEdit.tsx
@@ -5,7 +5,7 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import useAlert from '@/hooks/useAlert';
 import { passwordSchema, PutPasswordFormData } from '@/apis/auth/types';
 import { usePutPassword } from '@/apis/auth/queries';
-import { getErrorMessage } from '@/utils/errorMessage';
+import { getErrorMessage } from '@/utils/network/errorMessage';
 import { Input } from '@/components/ui/Field/Input';
 import { Card, CardTitle } from '@/components/ui/Card';
 import Button from '@/components/ui/Button';
