+++ src/components/profile/ProfileEdit.tsx
@@ -6,7 +6,7 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import useAlert from '@/hooks/useAlert';
 import { UpdateUserForm, updateUserFormSchema, User } from '@/apis/users/types';
 import { useCreateProfileImage, useUpdateUser } from '@/apis/users/queries';
-import { getErrorMessage } from '@/utils/errorMessage';
+import { getErrorMessage } from '@/utils/network/errorMessage';
 import { Input } from '@/components/ui/Field/Input';
 import { ImageUpload } from '@/components/ui/Field/ImageUpload';
 import Button from '@/components/ui/Button';
