+++ src/components/todo/DetailTodo.tsx
@@ -7,7 +7,7 @@ import useConfirm from '@/hooks/useConfirm';
 import { Card } from '@/apis/cards/types';
 import { useColumnsQuery } from '@/apis/columns/queries';
 import { useRemoveCard } from '@/apis/cards/queries';
-import { getErrorMessage } from '@/utils/errorMessage';
+import { getErrorMessage } from '@/utils/network/errorMessage';
 import { formatDate } from '@/utils/formatDate';
 import CommentSection from '@/components/todo/CommentSection';
 import { Modal, ModalContent, ModalFooter, ModalHandle, ModalHeader } from '@/components/ui/Modal';
@@ -15,7 +15,7 @@ import Avatar from '@/components/ui/Avatar';
 import TagChip from '@/components/ui/Chip/TagChip';
 import RoundChip from '@/components/ui/Chip/RoundChip';
 import Dropdown from '@/components/ui/Dropdown';
-import { DEFAULT_CARD_IMAGE_URL } from '@/constants/paths';
+import { DEFAULT_CARD_IMAGE_URL } from '@/constants/todo';
 import x_img from '@/assets/icons/x.svg';
 
 interface DetailTodoProps {
