+++ src/constants/todo.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_CARD_IMAGE_URL = 'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/task_image/12-5_45251_1739412796075.jpeg';
